First Upload to GitHub, version 0.0.5
This first upload of the build files contains what is believed to be the slimmest possible version of the game (as it currently exists).  This commit was made from inside the now-new virtual development machine.

Features as of now include a basic UI, clickable buttons, scrollable text boxes, text input fields, statistic display bars, and an introductory story.  A lot of work has gone into refactoring the code from v0.0.1 to v0.0.5.  Attributions for this game are still a TODO, but they continue clutter up my tabs.